Refactor component structure: move components to _components directory, remove unused components, and update layout metadata
