+++ src/app/Home/page.jsx
@@ -1,22 +1,22 @@
-import DesktopMessage from '../components/DesktopMessage';
-import MHomeCategory from '../components/MHomeCategory';
-import MostBookedServices from '../components/MostBookedServices';
-import PopularServices from '../components/PopularServices';
-import SearchBar from '../components/SearchBar';
+import DesktopMessage from '../_components/DesktopMessage';
+import MHomeCategory from '../_components/MHomeCategory';
+import MostBookedServices from '../_components/MostBookedServices';
+import PopularServices from '../_components/PopularServices';
+import SearchBar from '../_components/SearchBar';
 
 export default function Home() {
-  return (
-    <>
-      {/* Desktop Message - Only shown on desktop */}
-      <DesktopMessage />
+    return (
+        <>
+            {/* Desktop Message - Only shown on desktop */}
+            <DesktopMessage />
 
-      {/* Mobile Content - Only shown on mobile */}
-      <div className="md:hidden">
-        <SearchBar />
-        <MHomeCategory />
-        <PopularServices />
-        <MostBookedServices />
-      </div>
-    </>
-  );
+            {/* Mobile Content - Only shown on mobile */}
+            <div className="md:hidden">
+                <SearchBar />
+                <MHomeCategory />
+                <PopularServices />
+                <MostBookedServices />
+            </div>
+        </>
+    );
 }
